chore(refactor): create split seperation of concern
