+++ README.md
@@ -223,96 +223,9 @@ So mixed setups are fine for getting files across, but they're **not fully priva
 - **Want no relays at all?** Both people set **Disabled** — only works when a direct connection is possible (e.g. same network).
 
 
+## Development
 
-## Development Setup
-
-### Prerequisites
-
-- Rust 1.91+
-- Node.js 20+
-- pnpm 10+
-
-### Getting Started
-
-1. **Fork and clone the repository**:
-   ```bash
-   git clone https://github.com/your-username/alt-sendme.git
-   cd alt-sendme
-   ```
-
-2. **Install dependencies**:
-   ```bash
-   pnpm install
-   ```
-
-3. **Install Tauri CLI** (desktop/Android):
-   ```bash
-   cargo install tauri-cli
-   ```
-
-### Desktop (Tauri)
-
-```bash
-pnpm tauri dev
-```
-
-Build for Tauri:
-
-```bash
-pnpm build
-pnpm tauri build
-```
-
-### Web preview (browser UI shell)
-
-Runs the React UI in the browser with Tauri APIs stubbed — transfers are not available yet.
-
-```bash
-pnpm dev:web
-```
-
-Open **`http://localhost:3001/web/`** (web mode uses the `/web/` base path for future embed at `altsendme.com/web`).
-
-Production static build:
-
-```bash
-pnpm build:web
-pnpm preview:web   # serves frontend/dist at http://localhost:3001/web/
-```
-
-### Android
-
-1. **(Optional) Setup android project**:
-   ```bash
-   rm src-tauri/gen/android
-   cargo tauri android init
-   git checkout src-tauri/gen/android
-   pnpm android:dev
-   ```
-
-2. **Build locally**:
-   ```bash
-   cargo tauri build --no-bundle
-   ```
-
-3. **Install debug APK**:
-   ```bash
-   pnpm android:build -- --debug --apk
-   adb install -r src-tauri/gen/android/app/build/outputs/apk/universal/debug/app-universal-debug.apk
-   ```
-
-### Repo layout
-
-| Path | Purpose |
-|------|---------|
-| `frontend/` | React UI (Tauri + web targets) |
-| `src-tauri/` | Tauri desktop/Android shell |
-| `engine/` | Rust P2P transfer engine |
-| `www/` | Public Next.js site (planned — embeds web UI at `/web`) |
-
-## Testing Locally
-
-Install [Sendme CLI](https://www.iroh.computer/sendme) tool and you can share files within same device to test the whole transfer process. Files don't leave your device it works like a copy operation.
+See [CONTRIBUTING.md](CONTRIBUTING.md#development-setup) for prerequisites, local setup, build instructions, and testing.
 
 ## Join our [Discord](https://discord.gg/xwb7z22Eve) to contribute
