+++ engine/Cargo.toml
@@ -1,3 +1,7 @@
+[workspace]
+members = [".", "protocol", "native", "wasm-io"]
+resolver = "2"
+
 [package]
 name = "engine"
 version = "0.1.0"
@@ -9,53 +13,29 @@ license = "AGPL-3.0"
 repository = "https://github.com/tonyantony300/alt-sendme"
 description = "AltSendme - Frictionless peer-to-peer file transfer desktop application with NAT hole punching"
 readme = "README.md"
-
-# Sadly this also needs to be updated in .github/workflows/ci.yml
 rust-version = "1.91"
-
 
 [lib]
 name = "engine"
 crate-type = ["lib", "staticlib"]
 
 [dependencies]
-anyhow = "1.0.75"
-# Keep transitive Iroh stack crates above audited vulnerable versions.
-bytes = "=1.11.1"
-derive_more = { version = "2.0.1", features = ["display", "from_str"] }
-# I had some issues with futures-buffered 0.2.9
-futures-buffered = "0.2"
-rand = "0.10"
-rustls-webpki = "=0.103.13"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1.0.108"
-tracing = "0.1.40"
-data-encoding = "2.8.0"
-n0-future = "0.3"
-hex = "0.4.3"
+protocol = { path = "protocol", package = "sendme-protocol" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-iroh-blobs = { version = "0.103" }
-iroh = "1.0.0"
-tokio = { version = "1.34.0", features = ["full"] }
-tracing-subscriber = { version = "0.3.20", features = ["env-filter"] }
+native = { path = "native", package = "sendme-native" }
 clap = { version = "4.4.10", features = ["derive"] }
 console = "0.15.7"
 crossterm = { version = "0.29.0", features = [
   "event-stream",
   "osc52",
 ], optional = true }
-dirs = "6.0.0"
 indicatif = "0.17.7"
-num_cpus = "1.16.0"
-walkdir = "2.4.0"
+tracing-subscriber = { version = "0.3.20", features = ["env-filter"] }
+irpc = "0.17.0"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom = { version = "0.3", features = ["wasm_js"] }
-iroh-blobs = { version = "0.103", default-features = false }
-iroh = { version = "1.0.0", default-features = false, features = ["tls-ring"] }
-tokio = { version = "1.34.0", features = ["macros", "rt", "sync"] }
-web-time = "1"
+wasm-io = { path = "wasm-io", package = "sendme-wasm-io" }
 
 [target.'cfg(unix)'.dependencies]
 libc = { version = "0.2.174", optional = true }
@@ -71,6 +51,9 @@ nix = { version = "0.29", features = ["signal", "process"] }
 rand = "0.10"
 serde_json = "1.0.108"
 tempfile = "3.8.1"
+tokio = { version = "1.34.0", features = ["full"] }
+iroh-blobs = { version = "0.103" }
+data-encoding = "2.8.0"
 
 [features]
 clipboard = ["dep:crossterm", "dep:windows-sys", "dep:libc"]
