+++ engine/native/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "sendme-native"
+version = "0.1.0"
+edition = "2021"
+authors = ["Tony Antony <tonyantony300@gmail.com>"]
+license = "AGPL-3.0"
+repository = "https://github.com/tonyantony300/alt-sendme"
+description = "Native filesystem I/O for alt-sendme (desktop/mobile)"
+rust-version = "1.91"
+
+[lib]
+name = "native"
+
+[dependencies]
+protocol = { path = "../protocol", package = "sendme-protocol" }
+anyhow = "1.0.75"
+data-encoding = "2.8.0"
+dirs = "6.0.0"
+iroh-blobs = { version = "0.103" }
+iroh = "1.0.0"
+n0-future = "0.3"
+num_cpus = "1.16.0"
+rand = "0.10"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0.108"
+tokio = { version = "1.34.0", features = ["full"] }
+tracing = "0.1.40"
+walkdir = "2.4.0"
+
+[dev-dependencies]
+tempfile = "3.8.1"
