+++ engine/native/src/receive.rs
@@ -1,0 +1,172 @@
+use protocol::{
+    download_to_store, get_or_create_secret, AppHandle, ReceiveOptions,
+};
+use iroh::endpoint::presets;
+use iroh::{address_lookup::dns::DnsAddressLookup, Endpoint};
+use iroh_blobs::ticket::BlobTicket;
+use std::str::FromStr;
+use tokio::select;
+
+use crate::export::export_to_directory;
+use crate::storage;
+use crate::types::ReceiveResult;
+
+fn emit_event(app_handle: &AppHandle, event_name: &str) {
+    if let Some(handle) = app_handle {
+        if let Err(e) = handle.emit_event(event_name) {
+            tracing::warn!("Failed to emit event {}: {}", event_name, e);
+        }
+    }
+}
+
+fn emit_event_with_payload(app_handle: &AppHandle, event_name: &str, payload: &str) {
+    if let Some(handle) = app_handle {
+        if let Err(e) = handle.emit_event_with_payload(event_name, payload) {
+            tracing::warn!("Failed to emit event {} with payload: {}", event_name, e);
+        }
+    }
+}
+
+pub async fn download(
+    ticket_str: String,
+    options: ReceiveOptions,
+    app_handle: AppHandle,
+) -> anyhow::Result<ReceiveResult> {
+    let ticket = BlobTicket::from_str(&ticket_str)?;
+    let addr = ticket.addr().clone();
+    let secret_key = get_or_create_secret()?;
+
+    let mut builder = Endpoint::builder(presets::Minimal)
+        .alpns(vec![])
+        .secret_key(secret_key)
+        .relay_mode(options.relay_mode.clone().into());
+
+    if ticket.addr().relay_urls().count() == 0 && ticket.addr().ip_addrs().count() == 0 {
+        builder = builder.address_lookup(DnsAddressLookup::n0_dns());
+    }
+    if let Some(addr) = options.magic_ipv4_addr {
+        builder = builder.bind_addr(addr)?;
+    }
+    if let Some(addr) = options.magic_ipv6_addr {
+        builder = builder.bind_addr(addr)?;
+    }
+
+    let endpoint = builder.bind().await?;
+    let (db, iroh_data_dir) =
+        storage::create_recv_store(&ticket.hash().to_hex().to_string()).await?;
+    let mut cleanup_guard = storage::recv_cleanup_guard(iroh_data_dir);
+    let db2 = db.clone();
+
+    let fut = async move {
+        let downloaded =
+            download_to_store(ticket, addr, &endpoint, db.as_ref(), &app_handle).await?;
+
+        let output_dir = options.output_dir.unwrap_or_else(|| {
+            dirs::download_dir().unwrap_or_else(|| std::env::current_dir().unwrap())
+        });
+
+        let conflicts =
+            export_to_directory(&db, downloaded.collection, &output_dir).await?;
+
+        if !conflicts.is_empty() {
+            let payload = serde_json::to_string(&conflicts).unwrap_or_else(|_| "[]".to_string());
+            emit_event_with_payload(&app_handle, "receive-conflicts", &payload);
+        }
+
+        endpoint.close().await;
+        emit_event(&app_handle, "receive-completed");
+
+        anyhow::Ok((
+            downloaded.total_files,
+            downloaded.payload_size,
+            downloaded.stats,
+            output_dir,
+            conflicts.len(),
+        ))
+    };
+
+    let (total_files, payload_size, _stats, output_dir, conflict_count) = select! {
+        x = fut => match x {
+            Ok(x) => x,
+            Err(e) => {
+                tracing::error!("Download operation failed: {}", e);
+                cleanup_guard.disarm();
+                db2.shutdown().await?;
+                anyhow::bail!("error: {e}");
+            }
+        },
+        _ = tokio::signal::ctrl_c() => {
+            tracing::warn!("Operation cancelled by user");
+            db2.shutdown().await?;
+            anyhow::bail!("Operation cancelled");
+        }
+    };
+
+    let message = if conflict_count > 0 {
+        format!(
+            "Downloaded {} files, {} bytes ({} name conflicts auto-resolved)",
+            total_files, payload_size, conflict_count
+        )
+    } else {
+        format!("Downloaded {} files, {} bytes", total_files, payload_size)
+    };
+
+    Ok(ReceiveResult {
+        message,
+        file_path: output_dir,
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::send::start_share;
+    use protocol::{
+        AddrInfoOptions, FileMetadata, RelayModeOption, SendOptions,
+    };
+    use std::io::Write;
+    use tempfile::NamedTempFile;
+
+    #[tokio::test]
+    async fn test_fetch_metadata_e2e() {
+        let mut temp_file = NamedTempFile::new().unwrap();
+        write!(temp_file, "metadata e2e test content").unwrap();
+        let temp_path = temp_file.path().to_path_buf();
+
+        let expected_metadata = FileMetadata {
+            file_name: "test_e2e_file.txt".into(),
+            item_count: 1,
+            size: 25,
+            thumbnail: Some("data:image/jpeg;base64,e2e_test_thumbnail=".into()),
+            mime_type: Some("text/plain".into()),
+            items: None,
+        };
+
+        let send_opts = SendOptions {
+            relay_mode: RelayModeOption::Default,
+            ticket_type: AddrInfoOptions::RelayAndAddresses,
+            magic_ipv4_addr: None,
+            magic_ipv6_addr: None,
+        };
+
+        let result = start_share(temp_path, send_opts, None, Some(expected_metadata.clone()))
+            .await
+            .expect("Failed to start share");
+
+        let recv_opts = ReceiveOptions {
+            output_dir: None,
+            relay_mode: RelayModeOption::Default,
+            magic_ipv4_addr: None,
+            magic_ipv6_addr: None,
+        };
+
+        let fetched = protocol::fetch_metadata(result.ticket, recv_opts)
+            .await
+            .expect("Failed to fetch metadata from node");
+
+        assert_eq!(fetched.file_name, expected_metadata.file_name);
+        assert_eq!(fetched.size, expected_metadata.size);
+        assert_eq!(fetched.thumbnail, expected_metadata.thumbnail);
+        assert_eq!(fetched.mime_type, expected_metadata.mime_type);
+    }
+}
