+++ engine/protocol/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "sendme-protocol"
+version = "0.1.0"
+edition = "2021"
+authors = ["Tony Antony <tonyantony300@gmail.com>"]
+license = "AGPL-3.0"
+repository = "https://github.com/tonyantony300/alt-sendme"
+description = "Shared P2P transfer protocol for alt-sendme (iroh, metadata, progress)"
+rust-version = "1.91"
+
+[lib]
+name = "protocol"
+
+[dependencies]
+anyhow = "1.0.75"
+bytes = "=1.11.1"
+derive_more = { version = "2.0.1", features = ["display", "from_str"] }
+n0-future = "0.3"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0.108"
+tracing = "0.1.40"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+iroh-blobs = { version = "0.103" }
+iroh = "1.0.0"
+tokio = { version = "1.34.0", features = ["full"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { version = "0.3", features = ["wasm_js"] }
+iroh-blobs = { version = "0.103", default-features = false }
+iroh = { version = "1.0.0", default-features = false, features = ["tls-ring"] }
+tokio = { version = "1.34.0", features = ["io-util", "macros", "rt", "sync"] }
+web-time = "1"
