+++ engine/protocol/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod receive;
+pub mod send;
+pub mod time_compat;
+pub mod types;
+
+pub use receive::{download_to_store, fetch_metadata, DownloadToStoreResult};
+pub use send::{run_share_session, ShareSessionOutcome, METADATA_ALPN};
+pub use types::*;
+
+#[cfg(target_arch = "wasm32")]
+pub use types::set_wasm_secret_key;
