--- engine/src/core/.gitignore
@@ -1,2 +1,0 @@
-send_reference
-receive_reference
