+++ engine/wasm-io/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sendme-wasm-io"
+version = "0.1.0"
+edition = "2021"
+authors = ["Tony Antony <tonyantony300@gmail.com>"]
+license = "AGPL-3.0"
+repository = "https://github.com/tonyantony300/alt-sendme"
+description = "Browser in-memory I/O for alt-sendme (wasm32)"
+rust-version = "1.91"
+
+[lib]
+name = "wasm_io"
+
+[dependencies]
+protocol = { path = "../protocol", package = "sendme-protocol" }
+anyhow = "1.0.75"
+getrandom = { version = "0.3", features = ["wasm_js"] }
+iroh-blobs = { version = "0.103", default-features = false }
+iroh = { version = "1.0.0", default-features = false, features = ["tls-ring"] }
+n0-future = "0.3"
+tokio = { version = "1.34.0", features = ["macros", "rt", "sync"] }
+tracing = "0.1.40"
