+++ engine/wasm-io/src/lib.rs
@@ -1,0 +1,15 @@
+pub mod export;
+pub mod import;
+pub mod receive;
+pub mod send;
+pub mod storage;
+pub mod types;
+
+pub use protocol::{
+    apply_options, fetch_metadata, get_or_create_secret, set_wasm_secret_key, AddrInfoOptions,
+    AppHandle, EventEmitter, FileMetadata, FilePreviewItem, ReceiveOptions, RelayModeOption,
+    SendOptions,
+};
+pub use receive::download_bytes;
+pub use send::start_share_bytes;
+pub use types::{WasmReceiveResult, WasmShareSession};
