+++ engine/wasm-io/src/send.rs
@@ -1,0 +1,76 @@
+use protocol::{
+    get_or_create_secret, run_share_session, AddrInfoOptions, AppHandle, FileMetadata, SendOptions,
+    METADATA_ALPN,
+};
+use iroh::endpoint::presets;
+use iroh::{endpoint::RelayMode, Endpoint};
+use iroh_blobs::{
+    provider::events::{ConnectMode, EventMask, EventSender, RequestMode},
+    BlobsProtocol,
+};
+use tokio::sync::mpsc;
+
+use crate::import::import_single_file_bytes;
+use crate::storage::create_send_mem_store;
+use crate::types::WasmShareSession;
+
+pub async fn start_share_bytes(
+    file_name: String,
+    bytes: Vec<u8>,
+    options: SendOptions,
+    app_handle: &AppHandle,
+    metadata: Option<FileMetadata>,
+) -> anyhow::Result<WasmShareSession> {
+    let secret_key = get_or_create_secret()?;
+    let relay_mode: RelayMode = options.relay_mode.clone().into();
+    let ticket_type = AddrInfoOptions::Relay;
+
+    let builder = Endpoint::builder(presets::N0)
+        .alpns(vec![iroh_blobs::ALPN.to_vec(), METADATA_ALPN.to_vec()])
+        .secret_key(secret_key)
+        .relay_mode(relay_mode.clone());
+
+    let (progress_tx, progress_rx) = mpsc::channel(64);
+    let endpoint = builder.bind().await?;
+    let store = create_send_mem_store();
+
+    let blobs = BlobsProtocol::new(
+        &store,
+        Some(EventSender::new(
+            progress_tx,
+            EventMask {
+                connected: ConnectMode::Notify,
+                get: RequestMode::NotifyLog,
+                ..EventMask::DEFAULT
+            },
+        )),
+    );
+
+    let (temp_tag, size) = import_single_file_bytes(file_name, bytes, blobs.store()).await?;
+
+    let outcome = run_share_session(
+        endpoint,
+        store,
+        blobs,
+        temp_tag,
+        size,
+        metadata,
+        ticket_type,
+        app_handle,
+        "file".to_string(),
+        relay_mode,
+        None,
+        progress_rx,
+    )
+    .await?;
+
+    Ok(WasmShareSession {
+        ticket: outcome.ticket,
+        hash: outcome.hash,
+        size: outcome.size,
+        router: outcome.router,
+        temp_tag: outcome.temp_tag,
+        store: outcome.store,
+        _progress_handle: outcome.progress_handle,
+    })
+}
